+++ test/helper/schemas.js
@@ -2,8 +2,9 @@ import AsyncTestUtil from 'async-test-util';
 
 export const human = {
     title: 'human schema',
-    version: 0,
     description: 'describes a human being',
+    version: 0,
+    disableKeyCompression: false,
     type: 'object',
     properties: {
         passportId: {
@@ -30,6 +31,7 @@ export const humanDefault = {
     title: 'human schema',
     version: 0,
     description: 'describes a human being',
+    disableKeyCompression: false,
     type: 'object',
     properties: {
         passportId: {
@@ -56,6 +58,7 @@ export const humanDefault = {
 export const humanFinal = {
     title: 'human schema with age set final',
     version: 0,
+    disableKeyCompression: false,
     type: 'object',
     properties: {
         passportId: {
@@ -80,6 +83,7 @@ export const humanFinal = {
 export const simpleHuman = {
     title: 'human schema',
     version: 0,
+    disableKeyCompression: false,
     description: 'describes a simple human being',
     type: 'object',
     properties: {
@@ -97,6 +101,7 @@ export const simpleHuman = {
 export const simpleHumanV3 = {
     title: 'human schema',
     version: 3,
+    disableKeyCompression: false,
     description: 'describes a simple human being',
     type: 'object',
     properties: {
@@ -111,11 +116,10 @@ export const simpleHumanV3 = {
     required: ['passportId', 'age']
 };
 
-
-
 export const humanAgeIndex = {
     title: 'human schema',
     version: 0,
+    disableKeyCompression: false,
     description: 'describes a human being',
     type: 'object',
     properties: {
@@ -144,6 +148,7 @@ export const humanSubIndex = {
     title: 'human schema',
     version: 0,
     description: 'describes a human being where other.age is index',
+    disableKeyCompression: false,
     type: 'object',
     properties: {
         passportId: {
@@ -169,6 +174,7 @@ export const nestedHuman = {
     title: 'human nested',
     version: 0,
     description: 'describes a human being with a nested field',
+    disableKeyCompression: false,
     type: 'object',
     properties: {
         passportId: {
@@ -199,6 +205,7 @@ export const nestedHuman = {
 export const deepNestedHuman = {
     title: 'deep human nested',
     version: 0,
+    disableKeyCompression: false,
     description: 'describes a human being with a nested field',
     type: 'object',
     properties: {
@@ -234,6 +241,7 @@ export const noindexHuman = {
     title: 'human schema',
     version: 0,
     description: 'this schema has no index',
+    disableKeyCompression: false,
     type: 'object',
     properties: {
         firstName: {
@@ -250,6 +258,7 @@ export const noindexHuman = {
 export const nostringIndex = {
     description: 'the index has no type:string',
     version: 0,
+    disableKeyCompression: false,
     type: 'object',
     properties: {
         passportId: {
@@ -270,6 +279,7 @@ export const bigHuman = {
     title: 'human schema',
     version: 0,
     description: 'describes a human being with 2 indexes',
+    disableKeyCompression: false,
     type: 'object',
     properties: {
         passportId: {
@@ -300,6 +310,7 @@ export const encryptedHuman = {
     version: 0,
     description: 'uses an encrypted field',
     type: 'object',
+    disableKeyCompression: false,
     properties: {
         passportId: {
             type: 'string',
@@ -319,6 +330,7 @@ export const encryptedHuman = {
 export const encryptedObjectHuman = {
     title: 'human encrypted',
     version: 0,
+    disableKeyCompression: false,
     description: 'uses an encrypted field',
     type: 'object',
     properties: {
@@ -348,6 +360,7 @@ export const encryptedObjectHuman = {
 export const encryptedDeepHuman = {
     title: 'human encrypted',
     version: 0,
+    disableKeyCompression: false,
     description: 'uses an encrypted field',
     type: 'object',
     properties: {
@@ -411,6 +424,7 @@ export const compoundIndex = {
     version: 0,
     description: 'this schema has a compoundIndex',
     type: 'object',
+    disableKeyCompression: false,
     properties: {
         passportId: {
             type: 'string'
@@ -431,6 +445,7 @@ export const compoundIndexNoString = {
     title: 'compund index',
     version: 0,
     description: 'this schema has a compoundIndex',
+    disableKeyCompression: false,
     type: 'object',
     properties: {
         passportId: {
@@ -453,6 +468,7 @@ export const wrongCompoundFormat = {
     title: 'compund index',
     version: 0,
     description: 'this schema has a compoundIndex',
+    disableKeyCompression: false,
     type: 'object',
     properties: {
         passportId: {
@@ -482,6 +498,7 @@ export const empty = {
 export const heroArray = {
     'title': 'hero schema',
     version: 0,
+    disableKeyCompression: false,
     'description': 'describes a hero with an array-field',
     'type': 'object',
     'properties': {
@@ -513,6 +530,7 @@ export const simpleArrayHero = {
     title: 'hero schema',
     version: 0,
     description: 'describes a hero with a string-array-field',
+    disableKeyCompression: false,
     type: 'object',
     properties: {
         name: {
@@ -535,6 +553,7 @@ export const primaryHuman = {
     title: 'human schema with primary',
     version: 0,
     description: 'describes a human being with passsportID as primary',
+    disableKeyCompression: false,
     type: 'object',
     properties: {
         passportId: {
@@ -560,6 +579,7 @@ export const primaryHuman = {
 export const humanNormalizeSchema1 = {
     title: 'human schema',
     version: 0,
+    disableKeyCompression: false,
     description: 'describes a human being',
     properties: {
         age: {
@@ -575,6 +595,7 @@ export const humanNormalizeSchema1 = {
 export const humanNormalizeSchema2 = {
     title: 'human schema',
     version: 0,
+    disableKeyCompression: false,
     properties: {
         age: {
             minimum: 0,
@@ -591,6 +612,7 @@ export const humanNormalizeSchema2 = {
 export const refHuman = {
     title: 'human related to other human',
     version: 0,
+    disableKeyCompression: false,
     properties: {
         name: {
             primary: true,
@@ -606,6 +628,7 @@ export const refHuman = {
 export const refHumanNested = {
     title: 'human related to other human',
     version: 0,
+    disableKeyCompression: false,
     properties: {
         name: {
             primary: true,
@@ -631,6 +654,7 @@ export const averageSchema = () => ({
     title: 'averageSchema_' + AsyncTestUtil.randomString(5), // randomisation used so hash differs
     version: 0,
     type: 'object',
+    disableKeyCompression: false,
     properties: {
         id: {
             type: 'string',