+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 Breaking:
   - Upgraded to [pouchdb 7.0.0](https://pouchdb.com/2018/06/21/pouchdb-7.0.0.html)
-
+  - `disableKeyCompression` is set to `false` by default.
 
 Other:
   - Upgrade to eslint 5 (no more babel-eslint)